requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
sacrebleu==1.5.0
numpy==1.23.3
torch==1.13.0
torchaudio==0.13.0
tqdm==4.62.3
wandb==0.13.4
logzero==1.7.0
pandas==2.1.4
transformers==4.36.1
scikit-learn==1.3.2
soundfile==0.12.1
sentencepiece==0.1.99
nltk==3.8.1
bert_score==0.3.13